requirements.txt
Flask==2.3.3
joblib==1.4.2
scikit-learn==1.4.2
numpy==1.26.4
python-whois==0.9.5
urllib3==2.2.1
